+++ R/akin-deprecated.R
@@ -1,0 +1,11 @@
+#'@title Deprecated Functions In Package akin
+#'
+#'@description The following functions are deprecated. Where available, alternative functions are suggested
+#' at \code{help("common-deprecated")}.
+#' @name akin-deprecated
+#' @keywords internal
+#'
+#'@section \code{common}:
+#'Instead of \code{common}, use [fcommon]
+#'
+NULL