+++ R/akin-package.R
@@ -20,5 +20,7 @@
 #' @keywords internal
 #'
  "_PACKAGE"
+
  NULL
+