+++ R/zzz.R
@@ -9,11 +9,16 @@
 #' @importFrom callr r_bg
 #' @importFrom Matrix Matrix rowSums rsparsematrix
 #' @importFrom listenv listenv
-#' @importFrom future.apply future_sapply future_lapply
+#' @importFrom future.apply future_lapply
 #'
 
-    .onLoad = function(libname, pkgname) {
+     .onLoad = function(libname, pkgname) {
                          Mmatrix = getFromNamespace('Mmatrix', 'Matrix')
                    rsparsematrix = getFromNamespace('rsparsematrix', 'Matrix')
-                   requireNamespace("future", quietly = TRUE)
-                }
+
+                  }
+   .onAttach = function(libname, pkgname) {
+                      suppressMessages(
+                                   requireNamespace("future", quietly = TRUE)
+                                )
+                  }